traitement ewkb

+++ pyetl/outils/gestion_logs.py
@@ -9,19 +9,13 @@ LOGGER = logging.getLogger("pyetl")
 
 def getlog(args):
     """recherche s il y a une demande de fichier log dans les arguments"""
-    log = None
-    log_level = None
-    log_print = None
+    loginfo = dict()
     if args:
         for i in args:
-            if "log_file=" in i:
-                log = i.split("=")[1]
-            if "log_level=" in i:
-                log_level = i.split("=")[1]
-            if "log_print=" in i:
-                log_print = i.split("=")[1]
+            if "log_file=" in i or "log_level=" in i or "log_print=" in i:
+                loginfo.update(i.split("=")[1])
     # print("dans getlog", log, log_level, log_print)
-    return log, log_level, log_print
+    return loginfo
 
 
 class DiffLogFilTer(logging.Filter):
@@ -74,7 +68,7 @@ class GestionLogs(object):
         }
         self.niveau_p = "INFO"
         self.log_level = "DEBUG"
-        self.niveau_f = self.log_level
+        self.niveau_f = "DEBUG"
         self.cur_mapper = mapper
         self.fichier = None
 
@@ -105,7 +99,7 @@ class GestionLogs(object):
         if self.aff_handler:
             self.logger.addHandler(self.aff_handler)
 
-    def configure_print_handlers(self):
+    def configure_print_handlers(self, niveau_p):
         """ genere les configs d affichage """
         self.resetlog()
         # self.set_logs()
@@ -117,29 +111,33 @@ class GestionLogs(object):
         self.print_handler.prec_record = ""
         self.print_handler.addFilter(lambda x: x.levelno != 999)
         self.print_handler.addFilter(DiffLogFilTer())
-        self.print_handler.setLevel(self.niveau_p)
+        self.print_handler.setLevel(niveau_p)
         aff_formatter = logging.Formatter("========================== %(message)s")
         self.aff_handler.setLevel(self.loglevels["AFFICH"])
         self.aff_handler.setFormatter(aff_formatter)
         self.add_handlers()
 
-    def resetlogfile(self, mode):
+    def resetlogfile(self, mode, level=""):
+        if self.cur_mapper.worker:
+            return  # on ne fait pas en mode parrallele
         if self.file_handler:
+            if mode == "stop":
+                self.logger.info("arret log en fichier")
             self.file_handler.close()
             self.logger.removeHandler(self.file_handler)
             if mode == "del":
                 os.remove(self.fichier)
-                print("resetlog", self.fichier)
-                self.setlogfile()
+                # print("resetlog", self.fichier)
+                self.setlogfile(level)
             elif mode == "rotate_w":
                 jour = time.strftime("%w")
                 fich, ext = os.path.splitext(self.fichier)
                 fich = fich + "_" + jour
                 self.fichier = fich + "." + ext
                 os.remove(self.fichier)
-                self.setlogfile()
+                self.setlogfile(level)
 
-    def setlogfile(self):
+    def setlogfile(self, level=""):
         fichier = self.fichier
         os.makedirs(os.path.dirname(fichier), exist_ok=True)
         fileformatter = logging.Formatter(
@@ -151,9 +149,11 @@ class GestionLogs(object):
         self.file_handler = logging.FileHandler(fichier)
         # on lui met le niveau sur DEBUG, on lui dit qu'il doit utiliser le formateur
         # créé précédement et on ajoute ce handler au logger
+        if level in self.loglevels:
+            self.niveau_f = level
         self.file_handler.setLevel(self.niveau_f)
         self.file_handler.setFormatter(fileformatter)
-        print("ajout filehandler", self.fichier, self.niveau_f)
+        # print("ajout filehandler", self.fichier, self.niveau_f)
         self.logger.addHandler(self.file_handler)
         self.logger.log(999, "demarrage pyetl %s", self.cur_mapper.getvar("VERSION"))
         self.logger.info("commande:   %s", self.cur_mapper.getvar("pyetl_script_ref"))
@@ -161,25 +161,24 @@ class GestionLogs(object):
         if args:
             self.logger.info("parametres: %s", args)
 
-    def initlog(self, mapper, loginfo=None, force=False):
+    def initlog(self, mapper, force=False):
         """initialise le contexte de logging (parametres de site environnement)"""
         # if self.loginited and not force and mapper == self.cur_mapper:
         if self.loginited and not force:
             return  # on a deja fait le boulot
-        print("initlog", mapper.idpyetl, loginfo)
         self.cur_mapper = mapper
-        log_file = self.fichier
-        log_print = self.niveau_p
-        log_level = self.log_level
-        if loginfo:
-            logfile, loglevel, logprint = loginfo
-            log_file = logfile or log_file
-            log_level = loglevel or log_level
-            log_print = logprint or log_print
-        fichier = mapper.getvar("log_file", log_file)
-        log = mapper.getvar("log_level", log_level)
-        affich = mapper.getvar("log_print", log_print)
+
+        fichier = mapper.getvar("log_file", self.fichier)
+        log_level = mapper.getvar("log_level", self.log_level)
+        if log_level not in self.loglevels:
+            print("erreur log_level")
+            log_level = self.log_level
+        affich = mapper.getvar("log_print", self.niveau_p)
+        if affich not in self.loglevels:
+            print("erreur log_print")
+            affich = self.niveau_p
         logmode = mapper.getvar("log_mode", "prod")
+        # print("initlog", mapper.idpyetl, fichier, log_level, "print:", affich)
         logging.captureWarnings(True)
         self.loginited = True
 
@@ -194,15 +193,15 @@ class GestionLogs(object):
             logging.raiseExceptions = False
         worker = mapper.worker
 
-        self.niveau_f = self.loglevels.get(log, logging.INFO)
-        self.niveau_p = self.loglevels.get(affich, logging.ERROR)
-        self.loglevel = min(self.niveau_f, self.niveau_p)
+        niveau_f = self.loglevels.get(log_level)
+        niveau_p = self.loglevels.get(affich)
+
         # print ('niveaux de logging',niveau_f,niveau_p)
-        logger.setLevel(self.loglevel)
+        logger.setLevel(niveau_f)
 
         if not worker:
-            print("initialisation log", affich, log, "(", fichier, ")")
-            self.configure_print_handlers()
+            # print("initialisation log", affich, log_level, "(", fichier, ")")
+            self.configure_print_handlers(niveau_p)
             if fichier != self.fichier:
                 self.resetlogfile("stop")
                 self.fichier = fichier
@@ -213,14 +212,20 @@ class GestionLogs(object):
 
     def setdebug(self):
         """positionne le logger en mode debug pour l affichage d objets"""
-        self.logger.setLevel(min(self.loglevel, self.loglevels["DEBUG"]))
+        self.logger.setLevel(
+            min(self.loglevels[self.log_level], self.loglevels["DEBUG"])
+        )
         if self.print_handler:
-            self.print_handler.setLevel(min(self.niveau_p, self.loglevels["DEBUG"]))
+            self.print_handler.setLevel(
+                min(self.loglevels[self.niveau_p], self.loglevels["DEBUG"])
+            )
         if self.file_handler:
-            self.file_handler.setLevel(min(self.niveau_f, self.loglevels["DEBUG"]))
+            self.file_handler.setLevel(
+                min(self.loglevels[self.niveau_f], self.loglevels["DEBUG"])
+            )
 
     def stopdebug(self):
-        self.logger.setLevel(self.loglevel)
+        self.logger.setLevel(self.log_level)
         if self.print_handler:
             self.print_handler.setLevel(self.niveau_p)
         if self.file_handler:
